requirements.txt
altair==5.5.0
attrs==25.3.0
audioread==3.0.1
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
filelock==3.13.1
fonttools==4.58.4
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.4
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
msgpack==1.1.1
narwhals==1.44.0
networkx==3.3
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
protobuf==6.31.1
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
smmap==5.0.2
soundfile==0.13.1
soxr==0.5.0.post1
streamlit==1.46.0
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tornado==6.5.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
